# AVRO-2523: Perf's usage doesn't show the option for specific record test